Edit and Delete Note done in AR and Backend

+++ StickyNotes_iOS/StickyNotes/Networking/Data/StickyNoteColor.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct StickyNoteColor: Codable {
-    var red: Double
-    var green: Double
-    var blue: Double
+    var red: Float
+    var green: Float
+    var blue: Float
 }
